moved the update product form to admin dashboard
